+++ generators/storybook/options.go
@@ -1,0 +1,11 @@
+package storybook
+
+// Options for generating storybook
+type Options struct {
+	// add your stuff here
+}
+
+// Validate that options are usuable
+func (opts *Options) Validate() error {
+	return nil
+}