+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # CHANGELOG
 
 ## Unreleased
+
+- add `g storybook:install`
+- add config package
+- upgrade to apollo client v3